requirements.txt
torch>=2.0.0
torch-geometric>=2.3.0
torch-scatter
torch-sparse
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0
pyyaml>=6.0
entmax>=1.0          # Sparsemax implementation
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
gseapy>=1.0.0        # Gene Set Enrichment Analysis
